Insert comma to separate the last element
Rationale: the serial comma is the last comma before coordinating conjuction and the final item in a series. Its use is considered a matter of style and is cause for much debate.

There are times, however, it must be used to avoid confusion or ambiguity. For the sake of consistency, I recommend always using the serial comma.

In the series "mockups, assets and Symbols", one might erroneously think that there's a trash for two things: one for "mockups" and one for "assets and Symbols".